requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pyotp
qrcode[pil]
pycryptodome
python-dotenv
pydantic-settings